move utilities out of codegen package
