+++ hadoop-ozone/recon-codegen/src/main/java/org/apache/ozone/recon/schema/ReconSqlDbConfig.java
@@ -15,7 +15,7 @@
  * limitations under the License.
  */
 
-package org.apache.ozone.recon.codegen;
+package org.apache.ozone.recon.schema;
 
 import static java.util.concurrent.TimeUnit.SECONDS;
