+++ hadoop-ozone/recon-codegen/src/main/java/org/apache/ozone/recon/schema/SchemaVersionTableDefinition.java
@@ -17,8 +17,8 @@
 
 package org.apache.ozone.recon.schema;
 
-import static org.apache.ozone.recon.codegen.SqlDbUtils.TABLE_EXISTS_CHECK;
-import static org.apache.ozone.recon.codegen.SqlDbUtils.listAllTables;
+import static org.apache.ozone.recon.schema.SqlDbUtils.TABLE_EXISTS_CHECK;
+import static org.apache.ozone.recon.schema.SqlDbUtils.listAllTables;
 import static org.jooq.impl.DSL.name;
 
 import com.google.inject.Inject;
