+++ hadoop-ozone/recon-codegen/src/main/java/org/apache/ozone/recon/schema/SqlDbUtils.java
@@ -15,7 +15,7 @@
  * limitations under the License.
  */
 
-package org.apache.ozone.recon.codegen;
+package org.apache.ozone.recon.schema;
 
 import static org.jooq.impl.DSL.count;
