+++ hadoop-ozone/recon-codegen/src/main/java/org/apache/ozone/recon/schema/UtilizationSchemaDefinition.java
@@ -17,7 +17,7 @@
 
 package org.apache.ozone.recon.schema;
 
-import static org.apache.ozone.recon.codegen.SqlDbUtils.TABLE_EXISTS_CHECK;
+import static org.apache.ozone.recon.schema.SqlDbUtils.TABLE_EXISTS_CHECK;
 
 import com.google.inject.Inject;
 import com.google.inject.Singleton;
