+++ hadoop-ozone/recon/src/main/java/org/apache/hadoop/ozone/recon/persistence/DerbyDataSourceProvider.java
@@ -18,7 +18,7 @@
 package org.apache.hadoop.ozone.recon.persistence;
 
 import static org.apache.ozone.recon.codegen.JooqCodeGenerator.RECON_SCHEMA_NAME;
-import static org.apache.ozone.recon.codegen.SqlDbUtils.createNewDerbyDatabase;
+import static org.apache.ozone.recon.schema.SqlDbUtils.createNewDerbyDatabase;
 
 import com.google.inject.Inject;
 import com.google.inject.Provider;
